src/third_party/WebKit/LayoutTests/external/wpt/media-capabilities/idlharness.any-expected.txt

This is a testharness.js-based test.
Found 63 tests; 43 PASS, 20 FAIL, 0 TIMEOUT, 0 NOTRUN.
PASS idlharness
PASS idl_test setup
PASS Partial interface Navigator: original interface defined
PASS Partial interface Navigator: valid exposure set
PASS Partial interface WorkerNavigator: original interface defined
PASS Partial interface WorkerNavigator: valid exposure set
PASS Partial interface Screen: original interface defined
PASS MediaCapabilitiesInfo interface: existence and properties of interface object
PASS MediaCapabilitiesInfo interface object length
PASS MediaCapabilitiesInfo interface object name
PASS MediaCapabilitiesInfo interface: existence and properties of interface prototype object
PASS MediaCapabilitiesInfo interface: existence and properties of interface prototype object's "constructor" property
PASS MediaCapabilitiesInfo interface: existence and properties of interface prototype object's @@unscopables property
PASS MediaCapabilitiesInfo interface: attribute supported
PASS MediaCapabilitiesInfo interface: attribute smooth
PASS MediaCapabilitiesInfo interface: attribute powerEfficient
PASS MediaCapabilitiesInfo must be primary interface of decodingInfo
PASS Stringification of decodingInfo
PASS MediaCapabilitiesInfo interface: decodingInfo must inherit property "supported" with the proper type
PASS MediaCapabilitiesInfo interface: decodingInfo must inherit property "smooth" with the proper type
PASS MediaCapabilitiesInfo interface: decodingInfo must inherit property "powerEfficient" with the proper type
PASS MediaCapabilitiesInfo must be primary interface of encodingInfo
PASS Stringification of encodingInfo
PASS MediaCapabilitiesInfo interface: encodingInfo must inherit property "supported" with the proper type
PASS MediaCapabilitiesInfo interface: encodingInfo must inherit property "smooth" with the proper type
PASS MediaCapabilitiesInfo interface: encodingInfo must inherit property "powerEfficient" with the proper type
PASS MediaCapabilities interface: existence and properties of interface object
PASS MediaCapabilities interface object length
PASS MediaCapabilities interface object name
PASS MediaCapabilities interface: existence and properties of interface prototype object
PASS MediaCapabilities interface: existence and properties of interface prototype object's "constructor" property
PASS MediaCapabilities interface: existence and properties of interface prototype object's @@unscopables property
PASS MediaCapabilities interface: operation decodingInfo(MediaDecodingConfiguration)
PASS MediaCapabilities interface: operation encodingInfo(MediaEncodingConfiguration)
PASS MediaCapabilities must be primary interface of navigator.mediaCapabilities
PASS Stringification of navigator.mediaCapabilities
PASS MediaCapabilities interface: navigator.mediaCapabilities must inherit property "decodingInfo(MediaDecodingConfiguration)" with the proper type
PASS MediaCapabilities interface: calling decodingInfo(MediaDecodingConfiguration) on navigator.mediaCapabilities with too few arguments must throw TypeError
PASS MediaCapabilities interface: navigator.mediaCapabilities must inherit property "encodingInfo(MediaEncodingConfiguration)" with the proper type
PASS MediaCapabilities interface: calling encodingInfo(MediaEncodingConfiguration) on navigator.mediaCapabilities with too few arguments must throw TypeError
FAIL ScreenLuminance interface: existence and properties of interface object assert_own_property: self does not have own property "ScreenLuminance" expected property "ScreenLuminance" missing
FAIL ScreenLuminance interface object length assert_own_property: self does not have own property "ScreenLuminance" expected property "ScreenLuminance" missing
FAIL ScreenLuminance interface object name assert_own_property: self does not have own property "ScreenLuminance" expected property "ScreenLuminance" missing
FAIL ScreenLuminance interface: existence and properties of interface prototype object assert_own_property: self does not have own property "ScreenLuminance" expected property "ScreenLuminance" missing
FAIL ScreenLuminance interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "ScreenLuminance" expected property "ScreenLuminance" missing
FAIL ScreenLuminance interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "ScreenLuminance" expected property "ScreenLuminance" missing
FAIL ScreenLuminance interface: attribute min assert_own_property: self does not have own property "ScreenLuminance" expected property "ScreenLuminance" missing
FAIL ScreenLuminance interface: attribute max assert_own_property: self does not have own property "ScreenLuminance" expected property "ScreenLuminance" missing
FAIL ScreenLuminance interface: attribute maxAverage assert_own_property: self does not have own property "ScreenLuminance" expected property "ScreenLuminance" missing
FAIL ScreenLuminance must be primary interface of screen.luminance assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL Stringification of screen.luminance assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL ScreenLuminance interface: screen.luminance must inherit property "min" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL ScreenLuminance interface: screen.luminance must inherit property "max" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL ScreenLuminance interface: screen.luminance must inherit property "maxAverage" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
PASS Navigator interface: attribute mediaCapabilities
PASS Navigator interface: navigator must inherit property "mediaCapabilities" with the proper type
PASS WorkerNavigator interface: existence and properties of interface object
FAIL Screen interface: attribute colorGamut assert_true: The prototype object must have a property "colorGamut" expected true got false
FAIL Screen interface: attribute luminance assert_true: The prototype object must have a property "luminance" expected true got false
FAIL Screen interface: attribute onchange assert_true: The prototype object must have a property "onchange" expected true got false
FAIL Screen interface: screen must inherit property "colorGamut" with the proper type assert_inherits: property "colorGamut" not found in prototype chain
FAIL Screen interface: screen must inherit property "luminance" with the proper type assert_inherits: property "luminance" not found in prototype chain
FAIL Screen interface: screen must inherit property "onchange" with the proper type assert_inherits: property "onchange" not found in prototype chain
Harness: the test ran to completion.
